ADD SQL files for judge_libpq.c
judge_libpq.cで使用するSQL書いたファイルを追加。

+++ postgresql/table_create/users.sql
@@ -9,10 +9,10 @@ CREATE TABLE users(
 	adress_code varchar(255), -- 住所
 	email_adress varchar(100), --　email
 	birthday_date date, -- 生年月日
-	user_level integer(1), -- ユーザレベル
-	roll integer(1), --　役職
-	department integer(1), --　学部
-	major integer(1), --　学科、プログラム
+	user_level integer, -- ユーザレベル
+	roll integer, --　役職
+	department integer, --　学部
+	major integer, --　学科、プログラム
 	school_year integer, --　学年
         PRIMARY KEY (rogin_id, id)
 );
